refactor(article): ♻️ Improved navigation encapsulation & type definitions.
    -	Encapsulated the `NavigationButton` component within
    `article/Head.vue` improving modularity.

    -	Updated the `NavigationNavMenu` implementation to adhere to
    recent changes in the `NavMenu` component.

    -	Renamed the `anchor` variable to `button` in `article/Head.vue`
    which improves naming accuracy.

    -	Updated all references to `mapblogPost` within the `[slug]` page
    and `pagination.ts` utility.

    -	Modified `navigation.d.ts` to make the `order` property optional,
    providing flexibility in navigation definitions.
