+++ frontend/pages/artikkel/[slug].vue
@@ -7,7 +7,7 @@
     //  --- Import dependencies & types
     import { useRoute } from 'vue-router';
     import { fetchCollection } from '#imports';
-    import { mapBlogData } from '@/composables/maps/blogPost';
+    import { mapBlogData } from '~/composables/maps/mapBlogPost';
 
     import type { DevPostsCollectionItem } from '@nuxt/content';
