+++ frontend/types/navigation.d.ts
@@ -2,7 +2,7 @@ import type { FigureItem } from '~/types/media';
 
 interface BaseItem { id?: string; label?: string; cls?: string[]; type?: string[]; }
 
-export interface RouterItem extends BaseItem { path: string; order: number; }
+export interface RouterItem extends BaseItem { path: string; order?: number; }
 export interface AnchorItem extends BaseItem { href: string; media?: FigureItem; isDownload?: boolean; isDisabled?: boolean;}
 export interface ButtonItem extends BaseItem { disabled?: boolean; anchor?: AnchorItem; action?: () =>  number | void; type?: "submit" | "reset"; }
